Modernize remaining master-backed endpoint clients
Rebuild the registered office address, filing history, officer appointments,
PSC list, charges, insolvency, and document endpoints against the live
Companies House APIs. This updates the prerelease models to match observed
payloads, removes legacy wire-enum assumptions that break on new values, and
moves the affected contracts onto the generated string-backed value-type
pattern where the API exposes stable enum-like strings. The work also promotes
registered office address to a public sub-client, adds dedicated URI builders
for appointments and insolvency, and expands document metadata handling for
real filenames, timestamps, and large content lengths.

Add the supporting generator inputs and serializer changes needed for the new
value types, including extra YAML groups for filing, charge, insolvency, and
PSC values plus converter support for single-string-or-array payloads backed by
record-struct value types. Update the affected tests across unit, scenario,
DI, source-generator, and integration suites with live-schema assertions for
09a-09g, and serialize the integration project's real API tests to avoid flaky
parallel execution against Companies House.

+++ src/CompaniesHouse/Response/Document/DocumentMetadataContentLength.cs
@@ -5,6 +5,6 @@ namespace CompaniesHouse.Response.Document
     public class DocumentMetadataContentLength
     {
         [JsonPropertyName("content_length")]
-        public int ContentLength { get; set; }
+        public long ContentLength { get; set; }
     }
 }
